Web layer - web layer is responsible for receiving request and returning the proper response. Everything in between is handled by the other layers, where the actual business logic happens. In Spring Boot, this layer is represented by controllers (annotated with @Controller or @RestController), named after the model-view-controller (MVC) pattern. We're creating the initial web layer for our RESTful Spring Boot webservice by defining a BankController. Additionally, we create our first integration test that covers all layers of the application to see if everything plugs together as it should. For this, we use Spring’s MockMvc for automated testing of REST APIs.

+++ src/main/kotlin/com/kotlinspring/KotlinWithSpringboot/model/Bank.kt
@@ -5,6 +5,6 @@ data class Bank(
     //when you declare a var or a val kotlin declares a getter for you
         val accountNumber: String,
         val trust: String,
-        val transactionFee: Double
+        val transactionFee: Int
     //Kotlin we use primary constructor which is declared up here
 )
